requirements.txt
arrayqueues==1.2.0b0
msgpack==0.6.1
numpy==1.16.4
pandas==0.25.0
pyodbc==4.0.22
python-dotenv==0.10.3
pyzmq==18.0.2
requests==2.22.0
SQLAlchemy==1.3.6
